Split interface of NS LCM SOL005.
