+++ MANO-API/src/main/java/com/ubiqube/etsi/mano/controller/nslcm/sol005/NsLcmSubscriptionsSol005Api.java
@@ -1,0 +1,81 @@
+package com.ubiqube.etsi.mano.controller.nslcm.sol005;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.ubiqube.etsi.mano.model.nslcm.sol005.SubscriptionsPost;
+import com.ubiqube.etsi.mano.model.nslcm.sol005.SubscriptionsPostQuery;
+
+@RestController
+public class NsLcmSubscriptionsSol005Api implements NsLcmSubscriptionsSol005 {
+
+	/**
+	 * Query multiple subscriptions.
+	 *
+	 * Query Subscription Information. The GET method queries the list of active
+	 * subscriptions of the functional block that invokes the method. It can be used
+	 * e.g. for resynchronization after error situations.
+	 *
+	 */
+	@Override
+	public ResponseEntity<List<Object>> subscriptionsGet(String accept) {
+		return new ResponseEntity<>(new ArrayList<>(), HttpStatus.NOT_IMPLEMENTED);
+	}
+
+	/**
+	 * Subscribe to NS lifecycle change notifications.
+	 *
+	 * The POST method creates a new subscription. This method shall support the URI
+	 * query parameters, request and response data structures, and response codes,
+	 * as specified in the Tables 6.4.16.3.1-1 and 6.4.16.3.1-2. Creation of two
+	 * subscription resources with the same callbackURI and the same filter can
+	 * result in performance degradation and will provide duplicates of
+	 * notifications to the OSS, and might make sense only in very rare use cases.
+	 * Consequently, the NFVO may either allow creating a subscription resource if
+	 * another subscription resource with the same filter and callbackUri already
+	 * exists (in which case it shall return the \&quot;201 Created\&quot; response
+	 * code), or may decide to not create a duplicate subscription resource (in
+	 * which case it shall return a \&quot;303 See Other\&quot; response code
+	 * referencing the existing subscription resource with the same filter and
+	 * callbackUri).
+	 *
+	 */
+	@Override
+	public ResponseEntity<SubscriptionsPost> subscriptionsPost(String accept, String contentType, SubscriptionsPostQuery body) {
+		// : Implement...
+		return null;
+	}
+
+	/**
+	 * Terminate a subscription.
+	 *
+	 * The DELETE method terminates an individual subscription. This method shall
+	 * support the URI query parameters, request and response data structures, and
+	 * response codes, as specified in the Tables 6.4.17.3.5-1 and 6.4.17.3.5-2.
+	 *
+	 */
+	@Override
+	public void subscriptionsSubscriptionIdDelete(String subscriptionId) {
+		// : Implement...
+	}
+
+	/**
+	 * Read an individual subscription resource.
+	 *
+	 * The GET method retrieves information about a subscription by reading an
+	 * individual subscription resource. This method shall support the URI query
+	 * parameters, request and response data structures, and response codes, as
+	 * specified in the Tables 6.4.17.3.2-1 and 6.4.17.3.2-2
+	 *
+	 */
+	@Override
+	public ResponseEntity<SubscriptionsPost> subscriptionsSubscriptionIdGet(String subscriptionId, String accept) {
+		// : Implement...
+		return null;
+	}
+
+}
