--- MANO-API/src/main/java/com/ubiqube/etsi/mano/controller/nslcm/sol005/SubscriptionsApi.java
@@ -1,5 +1,0 @@
-package com.ubiqube.etsi.mano.controller.nslcm.sol005;
-
-public class SubscriptionsApi {
-
-}
